Fixed notification scheduling

+++ selfie-server/controllers/activityController.ts
@@ -69,9 +69,14 @@ export const getActivityById = async (req: any, res: any) => {
 export const deleteActivity = async (req: any, res: any) => {
     const { id } = req.params;
     try {
-        await Activity.findByIdAndDelete(id);
-        await inviteController.deleteActivityInvites(id);
-        await jobSchedulerService.clearEventNotificationsById(id);
+
+        const activity = await Activity.findById(id);
+
+        if(activity){
+            await jobSchedulerService.clearActivityNotifications(activity);
+            await Activity.findByIdAndDelete(id);
+            await inviteController.deleteActivityInvites(id);
+        }
 
         res.status(204).send();
     } catch (error) {
